+++ scripts/RT/enqueue_weps2sweep_jobs.py
@@ -18,7 +18,7 @@ import click
 import httpx
 import pandas as pd
 import pika
-from enqueue_jobs import GRAPH_HUC12
+from enqueue_wepp_jobs import GRAPH_HUC12
 from pyiem.database import get_sqlalchemy_conn, sql_helper
 from pyiem.util import logger
